# bls-server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BLS-KOX API",
    "version": "1.0.0",
    "description": "BLS-KOX 多租户后台管理系统 API 文档\n\n## 认证\n在请求头添加 `Authorization: Bearer <token>`\n登录接口获取 token：`POST /api/auth/login`\n\n## 默认账号\n- 账号: `superadmin`  密码: `123456`  租户: `000000`\n- 账号: `admin`       密码: `123456`  租户: `100000`",
    "contact": {
      "name": "BLS-KOX",
      "url": "https://github.com/npcxl/BLS-KOX"
    }
  },
  "servers": [
    {
      "url": "http://localhost:6001",
      "description": "本地开发"
    },
    {
      "url": "/api",
      "description": "Nginx 反代（生产）"
    }
  ],
  "paths": {
...
    "/api/system/tenant/{tenantId}": {
      "get": {
        "tags": [
          "tenant"
        ],
        "summary": "tenant - 查询 :tenantId",
        "operationId": "get_api_system_tenant_tenantId_",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/tenant_get__api_system_tenant_tenantId_"
              }
            }
          }
        }
      }
    },
    "/api/system/tenant/add": {
      "post": {
        "tags": [
          "tenant"
        ],
        "summary": "tenant - 新增记录",
        "operationId": "post_api_system_tenant_add",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/tenant_post__api_system_tenant_add"
              }
            }
          }
        }
      }
    },
    "/api/system/tenant/edit": {
      "put": {
        "tags": [
          "tenant"
        ],
        "summary": "tenant - 修改记录",
        "operationId": "put_api_system_tenant_edit",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/tenant_put__api_system_tenant_edit"
              }
            }
          }
        }
      }
    },
    "/api/system/tenant/list": {
      "get": {
        "tags": [
          "tenant"
        ],
        "summary": "tenant - 分页列表",
        "operationId": "get_api_system_tenant_list",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          },
          {
            "name": "pageNum",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "pageNum"
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "pageSize"
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "keyword"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/tenant_get__api_system_tenant_list"
              }
            }
          }
        }
      }
    },
    "/api/system/tenant/public-list": {
      "get": {
        "tags": [
          "tenant"
        ],
        "summary": "tenant - 查询 public-list",
        "operationId": "get_api_system_tenant_public-list",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        }
      }
    },
    "/api/system/tenant/remove": {
      "delete": {
        "tags": [
          "tenant"
        ],
        "summary": "tenant - 删除记录",
        "operationId": "delete_api_system_tenant_remove",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/tenant_delete__api_system_tenant_remove"
              }
            }
          }
        }
      }
    },
    "/api/system/tenant/status": {
      "put": {
        "tags": [
          "tenant"
        ],
        "summary": "tenant - 切换状态",
        "operationId": "put_api_system_tenant_status",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "操作成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "响应数据"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "tenantId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/tenant_put__api_system_tenant_status"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "tenant_get__api_system_tenant_tenantId_": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          }
        }
      },
      "tenant_post__api_system_tenant_add": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          }
        }
      },
      "tenant_put__api_system_tenant_edit": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          }
        }
      },
      "tenant_get__api_system_tenant_list": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          }
        }
      },
      "tenant_delete__api_system_tenant_remove": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          }
        }
      },
      "tenant_put__api_system_tenant_status": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "tenantId"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录获取的 accessToken"
      }
    }
  },
...
}